Validated parameters and moved logic to services/
