+++ api/routes/__init__.py
@@ -1,5 +1,5 @@
 from flask import Blueprint, jsonify
-from api.routes.core.quiz_gen_api import core_quiz_gen_bp
+from api.routes.quiz_gen_api import core_quiz_gen_bp
 
 api_blueprint = Blueprint("API", __name__, url_prefix="/api/")
 api_blueprint.register_blueprint(core_quiz_gen_bp)
